requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
cryptography==42.0.8
waitress==3.0.2
Flask-Login==0.6.3
Werkzeug<3.1.0